pyproject.toml
[tool.poetry]
name = "bolice_bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
ImageHash = "^4.3.1"
Pillow = "^9.2.0"
Pyrogram = "^2.0.59"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
